chore: move assess password strength logic into the useGeneratePassword hooks to get rid of the unnecessary useEffect
