+++ src/hooks/useGeneratePassword.ts
@@ -14,6 +14,7 @@ const useGeneratePassword = () => {
   const [generatedPassword, setGeneratedPassword] = useState<
     string | undefined
   >(undefined)
+  const [passwordStrength, setPasswordStrength] = useState<number>(0)
   const crypto = window.crypto
 
   const numbers = "1234567890"
@@ -80,6 +81,7 @@ const useGeneratePassword = () => {
       .join("")
 
     setGeneratedPassword(shuffledPassword)
+    assessPasswordStrength(shuffledPassword)
   }, [
     enabledSettings,
     retrieveRandomCharacters,
@@ -90,7 +92,34 @@ const useGeneratePassword = () => {
     passwordCharacterLength,
   ])
 
-  return { password: generatedPassword, generatePassword }
+  const assessPasswordStrength = useCallback((password: string
+  ) => {
+    if (!password) {
+      return
+    }
+
+    const containsNumbers = /\d/.test(password)
+    const containsUpperCase = /[A-Z]/.test(password)
+    const containsLowerCase = /[a-z]/.test(password)
+    const containsSymbols = /[ `!@#$%^&*()_+\-=[\]{};':"\\|,.<>/?~]/.test(
+      password,
+    )
+
+    const passwordChecks = [
+      containsNumbers,
+      containsUpperCase,
+      containsLowerCase,
+      containsSymbols,
+    ]
+
+    setPasswordStrength(passwordChecks.filter(Boolean).length)
+  }, [])
+
+  return {
+    password: generatedPassword,
+    generatePassword,
+    passwordStrength,
+  }
 }
 
 export default useGeneratePassword
